requirements.txt
pyyaml
requests
responses
packaging
setuptools